Ports this branch onto the per-read ctx work from #10916, which retired
ConcurrentCmdCtx ctx ownership and gave executePlan a different signature.

- executePlan: take master's shape (no cmdCtx parameter, and the ctx-loan
  return on the AREQ_StartCursor failure path) and add this branch's spec_ref
  parameter in place of master's dummy_spec_ref, at the definition, the
  forward declaration and all three call sites.

- executeAggregateDeferred: same substitution in the cursor branch, keeping
  master's restructured replyCtx handling.

- Master had already added the AREQ_StartCursor failure handling that this
  branch makes reachable; noted the reachability at that site.

- Dropped this branch's DistAggregateCleanups header comment: it described
  ownership via ConcurrentCmdCtx_KeepRedisCtx, which master deleted along with
  CMDCTX_KEEP_RCTX. threadHandleCommand now frees the dispatch ctx
  unconditionally and the cleanup path returns the loan, which master documents
  in place.

+++ src/redisearch_rs/query_eval/src/nodes/fuzzy.rs
@@ -24,7 +24,7 @@ use crate::{
 
 /// The `apiVersion` from which a fuzzy expansion also covers the empty term.
 /// Below it the empty term is never expanded to, whatever the distance.
-const EMPTY_TERM_API_VERSION: u32 = 2;
+const EMPTY_TERM_API_VERSION: u8 = 2;
 
 /// `QN_FUZZY` — expand a token to every term within `max_dist` edits of it over
 /// the spec's terms trie, and union the per-term readers.
